split observer's functions
